\name{cubeStats-package}
\alias{cubeStats-package}
\alias{cubeStats}
\docType{package}
\title{
\packageTitle{cubeStats}
}
\description{
\packageDescription{cubeStats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cubeStats}
\packageIndices{cubeStats}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cubeStats}

Maintainer: \packageMaintainer{cubeStats}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}


