% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sliceAllNA}
\alias{sliceAllNA}
\title{Check if there are any Finite values in each slice}
\usage{
sliceAllNA(x, mis_val = -2147483648)
}
\arguments{
\item{x}{A 3d matrix/array.}

\item{mis_val}{An integer to use as the missing value if the input matrix
is an integer type. Argument is ignored if the input array is numeric.}
}
\value{
A vector of 1 and 0s. 1 indicates all values are NA
}
\description{
Check if there are any Finite values in each slice
}
\examples{
small_matrix <- array(1:625, c(5, 5, 5))
sliceAllNA(small_matrix)
}
