% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sliceEval}
\alias{sliceEval}
\title{Evaluate simple expressions}
\usage{
sliceEval(
  x,
  expression = c(">", "<", "==", "within"),
  value,
  na.rm = FALSE,
  mis_val = -2147483648
)
}
\arguments{
\item{x}{A 3d matrix/array.}

\item{expression}{Character: one of c(">", "<", "==", "within").}

\item{value}{Numeric Vector of length 1 in most cases or length 2 when 
`expression = "within"`}

\item{na.rm}{True or false, should NAs be removed before calculating the mean.}

\item{mis_val}{An integer to use as the missing value if the input matrix
is an integer type. Argument is ignored if the input array is numeric.}
}
\value{
A vector of layer/slice sums based on evaluation.
}
\description{
Evaluate simple expressions
}
\details{
Within uses x > value[1] && x < value[2]
}
\examples{
small_matrix <- array(1:625, c(5, 5, 5))
sliceSum(small_matrix)
}
