% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sliceMin}
\alias{sliceMin}
\title{Get the min value of each layer in a 3d cube}
\usage{
sliceMin(x, na.rm = FALSE, mis_val = -2147483648)
}
\arguments{
\item{x}{A 3d matrix/array.}

\item{na.rm}{True or false, should NAs be removed before calculating the mean.}

\item{mis_val}{An integer to use as the missing value if the input matrix
is an integer type. Argument is ignored if the input array is numeric.}
}
\value{
A vector of layer/slice minimums.
}
\description{
Get the min value of each layer in a 3d cube
}
\examples{
small_matrix <- array(1:625, c(5, 5, 5))
lyrMean(small_matrix)
}
